feat: Add DNS server version identification question
This commit adds a new question in the "Firewall and IDS/IPS Evasion - Medium Lab" section. The question asks the user to submit the DNS server version of the target after the configurations are transferred to the system. This will help our client assess the security of their DNS server.