```text feat: Add question about identifying the version of a running service
This commit adds a new question in the "Firewall and IDS/IPS Evasion - Hard Lab" section. The question asks the user to find out the version of a specific service mentioned by the client. The answer should be submitted as the flag. This addition will help our client assess the security of their network.